Move Account/AccountStore into report/account.rs
AccountStore was a bare type alias for DirectInternStore, so changing its
implementation meant changing its type. Turn it into a wrapper struct
(mirroring CommodityStore) and give Account its own module, consistent with
Commodity living in commodity.rs.
