+++ core/src/report/transaction.rs
@@ -1,7 +1,7 @@
 use chrono::NaiveDate;
 
 use super::{
-    context::Account,
+    account::Account,
     eval::{Amount, SingleAmount},
 };
